backend/pdf_thumbnail_service/requirements.txt
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
functions-framework==3.4.0
PyMuPDF==1.23.5
Pillow==10.0.0